+++ cloud-minecraft/cloud-velocity/src/main/java/cloud/commandframework/velocity/VelocityCaptionKeys.java
@@ -32,6 +32,7 @@ import java.util.LinkedList;
 
 /**
  * Velocity specific {@link Caption caption keys}
+ * @since 1.1.0
  */
 public final class VelocityCaptionKeys {
 
@@ -41,6 +42,11 @@ public final class VelocityCaptionKeys {
      * Variables: {input}
      */
     public static final Caption ARGUMENT_PARSE_FAILURE_PLAYER = of("argument.parse.failure.player");
+
+    /**
+     * Variables: {input}
+     */
+    public static final Caption ARGUMENT_PARSE_FAILURE_SERVER = of("argument.parse.failure.server");
 
     private VelocityCaptionKeys() {
     }